refactor: divide lib in submodules
