+++ src/shapepy/geometry/polygon.py
