+++ tests/bool2d/test_bool_infinite_intersect.py
@@ -4,18 +4,18 @@ This module tests when two shapes have common edges/segments
 
 import pytest
 
-from shapepy.primitive import Primitive
+from shapepy.bool2d.primitive import Primitive
 
 
 @pytest.mark.order(9)
 @pytest.mark.dependency(
     depends=[
-        "tests/test_polygon.py::test_end",
-        "tests/test_jordan_polygon.py::test_end",
-        "tests/test_jordan_curve.py::test_end",
-        "tests/test_primitive.py::test_end",
-        "tests/test_contains.py::test_end",
-        "tests/test_empty_whole.py::test_end",
+        "tests/geometry/test_polygon.py::test_end",
+        "tests/geometry/test_jordan_polygon.py::test_end",
+        "tests/geometry/test_jordan_curve.py::test_end",
+        "tests/bool2d/test_primitive.py::test_end",
+        "tests/bool2d/test_contains.py::test_end",
+        "tests/bool2d/test_empty_whole.py::test_end",
     ],
     scope="session",
 )
