+++ tests/bool2d/test_contains.py
@@ -8,18 +8,23 @@ from fractions import Fraction
 import numpy as np
 import pytest
 
-from shapepy.jordancurve import JordanCurve
-from shapepy.primitive import Primitive
-from shapepy.shape import ConnectedShape, DisjointShape, EmptyShape, WholeShape
+from shapepy.bool2d.primitive import Primitive
+from shapepy.bool2d.shape import (
+    ConnectedShape,
+    DisjointShape,
+    EmptyShape,
+    WholeShape,
+)
+from shapepy.geometry.jordancurve import JordanCurve
 
 
 @pytest.mark.order(7)
 @pytest.mark.dependency(
     depends=[
-        "tests/test_polygon.py::test_end",
-        "tests/test_jordan_polygon.py::test_end",
-        "tests/test_jordan_curve.py::test_end",
-        "tests/test_primitive.py::test_end",
+        "tests/geometry/test_polygon.py::test_end",
+        "tests/geometry/test_jordan_polygon.py::test_end",
+        "tests/geometry/test_jordan_curve.py::test_end",
+        "tests/bool2d/test_primitive.py::test_end",
     ],
     scope="session",
 )
