+++ tests/bool2d/test_empty_whole.py
@@ -7,18 +7,18 @@ from copy import copy
 
 import pytest
 
-from shapepy.primitive import Primitive
-from shapepy.shape import EmptyShape, WholeShape
+from shapepy.bool2d.primitive import Primitive
+from shapepy.bool2d.shape import EmptyShape, WholeShape
 
 
 @pytest.mark.order(8)
 @pytest.mark.dependency(
     depends=[
-        "tests/test_polygon.py::test_end",
-        "tests/test_jordan_polygon.py::test_end",
-        "tests/test_jordan_curve.py::test_end",
-        "tests/test_primitive.py::test_end",
-        "tests/test_contains.py::test_end",
+        "tests/geometry/test_polygon.py::test_end",
+        "tests/geometry/test_jordan_polygon.py::test_end",
+        "tests/geometry/test_jordan_curve.py::test_end",
+        "tests/bool2d/test_primitive.py::test_end",
+        "tests/bool2d/test_contains.py::test_end",
     ],
     scope="session",
 )
