+++ tests/bool2d/test_shape.py
@@ -7,20 +7,20 @@ import math
 
 import pytest
 
-from shapepy.jordancurve import JordanCurve
-from shapepy.primitive import Primitive
-from shapepy.shape import IntegrateShape, SimpleShape
+from shapepy.bool2d.primitive import Primitive
+from shapepy.bool2d.shape import IntegrateShape, SimpleShape
+from shapepy.geometry.jordancurve import JordanCurve
 
 
 @pytest.mark.order(8)
 @pytest.mark.dependency(
     depends=[
-        "tests/test_polygon.py::test_end",
-        "tests/test_jordan_polygon.py::test_end",
-        "tests/test_jordan_curve.py::test_end",
-        "tests/test_primitive.py::test_end",
-        "tests/test_contains.py::test_end",
-        "tests/test_empty_whole.py::test_end",
+        "tests/geometry/test_polygon.py::test_end",
+        "tests/geometry/test_jordan_polygon.py::test_end",
+        "tests/geometry/test_jordan_curve.py::test_end",
+        "tests/bool2d/test_primitive.py::test_end",
+        "tests/bool2d/test_contains.py::test_end",
+        "tests/bool2d/test_empty_whole.py::test_end",
     ],
     scope="session",
 )
