+++ tests/geometry/test_jordan_curve.py
@@ -9,15 +9,15 @@ import pynurbs
 import pytest
 
 from shapepy import Point2D
-from shapepy.jordancurve import JordanCurve
+from shapepy.geometry.jordancurve import JordanCurve
 
 
 @pytest.mark.order(6)
 @pytest.mark.dependency(
     depends=[
-        "tests/test_polygon.py::test_end",
-        "tests/test_curve.py::test_end",
-        "tests/test_jordan_polygon.py::test_end",
+        "tests/geometry/test_polygon.py::test_end",
+        "tests/geometry/test_curve.py::test_end",
+        "tests/geometry/test_jordan_polygon.py::test_end",
     ],
     scope="session",
 )
